Ban users who have an unpaid loan
The loansbot will ban users with an unpaid loan,
and unban them when the loan is marked paid. The
loansbot will not ban moderators. For additional
protection, the loansbot will no longer ever
interact with users with less than 1000 combined
karma or younger than 90 days old.

Add PaidSummonTests, which was missing.