No longer use regex parsing on summon responses
Adds the new package me.timothy.bots.summon.patterns which contains new
replacements for Pattern and Matcher from the regex pattern, except that
are tuned for the specific format that this bot tends to look for.

If the transition is smooth, the me.timothy.bots.summon.patterns package
will be moved to the Summonable Bot repository to allow it to be used in
other reddit bots.

Fixed a minor, rare bug that occurred during testing where it was possible
to get in a situation where the test database would not delete automatically
due to foreign key constraints

Various tests added to test new summon format.